requirements.txt
scikit-learn~=1.1.2
pandas~=1.5.1
pyyaml~=6.0
torch~=1.12.1
sentence-transformers~=2.2.2